+++ trolley/payment_gateway.py
@@ -3,7 +3,7 @@ import pprint
 from trolley.exceptions.invalidFieldException import InvalidFieldException
 import trolley.configuration
 
-from trolley.utils.meta import Meta
+from trolley.types.meta import Meta
 
 class PaymentGateway(object):
     """
